fixes #4 #valid_phone_number? test
The current test iterates over an array of phone numbers and expects calls to `#valid_phone_number` with each number to return `false`. The problem with this is that there are valid phone numbers in the array which would cause the test fail.
It should really be checking if any of the numbers in the array is invalid. So `valid_numbers.all?` instead of `valid_numbers.each` should be used.